Create a `quasar.contrib.matryoshka` package.
Move the scattered things that should be pushed up to Matryoshka here.
