+++ connector/src/main/scala/quasar/fs/QueryFile.scala
@@ -18,6 +18,7 @@ package quasar.fs
 
 import quasar.Predef._
 import quasar._, Planner._, RenderTree.ops._
+import quasar.contrib.matryoshka._
 import quasar.contrib.pathy._
 import quasar.effect.LiftedOps
 import quasar.fp._
