+++ connector/src/main/scala/quasar/qscript/QScriptCore.scala
@@ -18,6 +18,7 @@ package quasar.qscript
 
 import quasar.Predef._
 import quasar.RenderTree
+import quasar.contrib.matryoshka._
 import quasar.fp._
 
 import matryoshka._
