+++ connector/src/main/scala/quasar/qscript/TTypes.scala
@@ -17,7 +17,9 @@
 package quasar.qscript
 
 import quasar.Predef._
+import quasar.contrib.matryoshka._
 import quasar.fp._
+
 import matryoshka._
 import scalaz._, Scalaz._
