+++ connector/src/main/scala/quasar/qscript/Transform.scala
@@ -17,6 +17,7 @@
 package quasar.qscript
 
 import quasar._
+import quasar.contrib.matryoshka._
 import quasar.ejson.{Int => _, _}
 import quasar.fp._
 import quasar.qscript.MapFunc._
@@ -538,7 +539,7 @@ class Transform
       fromData(data).fold[PlannerError \/ MapFunc[T, FreeMap[T]]](
         {
           case Data.NA => Undefined[T, FreeMap[T]]().right
-          case d => NonRepresentableData(d).left
+          case d       => NonRepresentableData(d).left
         },
         Constant[T, FreeMap[T]](_).right) ∘ (mf =>
         EnvT((
