+++ foundation/src/main/scala/quasar/fp/tree/package.scala
@@ -17,6 +17,7 @@
 package quasar.fp
 
 import quasar.Predef._
+import quasar.contrib.matryoshka._
 
 import matryoshka._
 import scalaz._
