+++ marklogic/src/main/scala/quasar/physical/marklogic/qscript/QScriptCorePlanner.scala
@@ -18,7 +18,7 @@ package quasar.physical.marklogic.qscript
 
 import quasar.Predef.{Map => _, _}
 import quasar.NameGenerator
-import quasar.fp.ShowT
+import quasar.contrib.matryoshka.ShowT
 import quasar.physical.marklogic.xquery._
 import quasar.physical.marklogic.xquery.syntax._
 import quasar.qscript._
