+++ marklogic/src/main/scala/quasar/physical/marklogic/qscript/package.scala
@@ -16,11 +16,11 @@
 
 package quasar.physical.marklogic
 
-import quasar.SKI.κ
 import quasar.NameGenerator
-import quasar.fp.{freeCataM, interpretM, ShowT}
-import quasar.qscript._
+import quasar.SKI.κ
+import quasar.contrib.matryoshka.{freeCataM, interpretM, ShowT}
 import quasar.physical.marklogic.xquery.{PrologW, XQuery}
+import quasar.qscript._
 
 import matryoshka.Recursive
 import scalaz._, Scalaz._
