+++ mongodb/src/main/scala/quasar/physical/mongodb/bsoncodec.scala
@@ -17,6 +17,7 @@
 package quasar.physical.mongodb
 
 import quasar.Predef._
+import quasar.contrib.matryoshka._
 import quasar.ejson.EJson
 import quasar.fp._
 import quasar._, Planner._
